[ROCm] Fix hardcoded float8_e4m3fn in compressed_tensors and FBGEMM FP8
Use platform-aware fp8_dtype (e4m3fnuz on MI300X) for weight tensor
initialization in:
- compressed_tensors_w8a8_fp8.py: Weight parameter allocation
- compressed_tensors_w8a16_fp8.py: Weight parameter allocation
- fpgemm_fp8.py: FBGEMM FP8 weight parameter allocation

Without this fix, loading FP8-serialized models through compressed_tensors
or FBGEMM on AMD ROCm would create weights with the wrong FP8 dtype,
causing silent numerical errors or crashes.

+++ python/sglang/srt/layers/quantization/fpgemm_fp8.py
@@ -10,6 +10,7 @@ from torch.nn.parameter import Parameter
 
 from sglang.srt.layers.linear import LinearBase
 from sglang.srt.layers.parameter import ChannelQuantScaleParameter, ModelWeightParameter
+from sglang.srt.layers.quantization.fp8_kernel import fp8_dtype
 from sglang.srt.layers.quantization.base_config import (
     LinearMethodBase,
     QuantizationConfig,
@@ -127,7 +128,7 @@ class FBGEMMFp8LinearMethod(LinearMethodBase):
             data=torch.empty(
                 output_size_per_partition,
                 input_size_per_partition,
-                dtype=torch.float8_e4m3fn,
+                dtype=fp8_dtype,
             ),
             input_dim=1,
             output_dim=0,
